+++ src/test/groovy/fr/nereide/test/reference/TestReferenceInXml.groovy
@@ -113,4 +113,8 @@ class TestReferenceInXml extends BaseReferenceTestCase {
     void testUiLabelPropertyInTitleWithOtherString() {
         doTest(UiLabelReference.class, 'PasBaptiste')
     }
+
+    void testServiceReferenceInXmlDataFile() {
+        doTest(ServiceReference.class, 'Croissance3000Service')
+    }
 }